+++ .gitignore
@@ -2,3 +2,4 @@
 
 cmd.exe.lnk
 gitk.cmd.lnk
+/ProjectAddIn1/ProjectAddIn1/*.user
