--- ProjectAddIn1/ProjectAddIn1/ProjectAddIn1.csproj.user
@@ -1,8 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <PublishUrlHistory>C:\Repos\ProjectArasAddIn\</PublishUrlHistory>
-    <InstallUrlHistory />
-    <SupportUrlHistory />
-  </PropertyGroup>
-</Project>
