--- ProjectAddIn1/ProjectAddIn1/ProjectUploader.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MS_Project_Import_Export
-{
-    public class ProjectUploader
-    {
-
-    }
-}
